+++ tests/Nivara.Tests/IO/ParquetWriterTests.cs
@@ -230,7 +230,7 @@ public class ParquetWriterTests
     {
         // Arrange
         var nullableIntArray = new int?[] { 1, null, 3, null, 5 };
-        var intColumn = NivaraColumn<int>.CreateFromNullable(nullableIntArray);
+        var intColumn = NivaraColumn.CreateFromNullable(nullableIntArray);
 
         var stringArray = new string[] { "a", null!, "c", null!, "e" };
         var stringColumn = NivaraColumn<string>.CreateForReferenceType(stringArray);