+++ tests/Nivara.Tests/Query/WindowExpressionEvaluationTests.cs
@@ -127,7 +127,7 @@ public class WindowExpressionEvaluationTests
     [Test]
     public void Evaluate_RollingSum_OverNulls_PropagatesWindowSemantics()
     {
-        var input = Input(("A", NivaraColumn<int>.CreateFromNullable(new int?[] { 1, null, 3, 4 })));
+        var input = Input(("A", NivaraColumn.CreateFromNullable(new int?[] { 1, null, 3, 4 })));
         var fused = new FusedExpressionEvaluator();
 
         var result = fused.Evaluate(ColumnExpressions.RollingSum(ColumnExpressions.Col("A"), 2), input);
@@ -138,7 +138,7 @@ public class WindowExpressionEvaluationTests
     [Test]
     public void Evaluate_CumulativeCount_OverSourceWithNulls_CountsValidOnly()
     {
-        var input = Input(("A", NivaraColumn<int>.CreateFromNullable(new int?[] { 1, null, 3, 4 })));
+        var input = Input(("A", NivaraColumn.CreateFromNullable(new int?[] { 1, null, 3, 4 })));
         var fused = new FusedExpressionEvaluator();
 
         var result = fused.Evaluate(ColumnExpressions.CumulativeCount(ColumnExpressions.Col("A")), input);